+++ src/app.ts
@@ -2,7 +2,7 @@ import http from 'http'
 import path from 'path'
 import debugLib from 'debug'
 import express, { Request, Response, NextFunction, Application, RequestHandler, Router } from 'express'
-import { LaserController } from './routes'
+import { LaserController } from './controllers/LaserController'
 import { init as initSocket } from './socket'
 
 const debug = debugLib('laser:web')
